requirements.txt
openai==0.28.0
srt==3.5.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0